Moved several classes to their own files
